requirements.txt
# FastAPI Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Streamlit Frontend
streamlit==1.31.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# LLM Integrations
anthropic==0.18.1
openai==1.12.0
httpx==0.26.0

# Web Search & Scraping
beautifulsoup4==4.12.3
requests==2.31.0
playwright==1.41.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
aiofiles==23.2.1
asyncio==3.4.3
tenacity==8.2.3

# Logging
loguru==0.7.2
